mcp_server/requirements.txt
mcp
psycopg2-binary
chromadb
sentence-transformers
pydantic
